+++ tests/Aws/Tests/DynamoDb/DynamoDbClientTest.php
@@ -21,6 +21,7 @@ use Aws\Common\Signature\SignatureV4;
 use Aws\DynamoDb\DynamoDbClient;
 use Aws\DynamoDb\Enum\Type;
 use Guzzle\Common\Collection;
+use Guzzle\Service\Resource\Model;
 
 class DynamoDbClientTest extends \Guzzle\Tests\GuzzleTestCase
 {
@@ -79,6 +80,92 @@ class DynamoDbClientTest extends \Guzzle\Tests\GuzzleTestCase
         ));
 
         $this->assertSame($expected, $actual);
+    }
+
+    public function dataForGetItemsFromResultsTest()
+    {
+        return array(
+            // Scan/Query results
+            array(
+                array(
+                    'Items' => array(
+                        array('a' => array('S' => 'a1'), 'b' => array('S' => 'b1')),
+                        array('a' => array('S' => 'a2'), 'b' => array('S' => 'b2')),
+                    )
+                ),
+                array(
+                    array('a' => 'a1', 'b' => 'b1'),
+                    array('a' => 'a2', 'b' => 'b2'),
+                )
+            ),
+
+            // GetItem results
+            array(
+                array(
+                    'Item' => array(
+                        'a' => array('S' => 'a3'),
+                        'b' => array('S' => 'b3')
+                    ),
+                ),
+                array(
+                    array('a' => 'a3', 'b' => 'b3')
+                )
+            ),
+
+            // UpdateItem/PutItem results
+            array(
+                array(
+                    'Attributes' => array(
+                        'a' => array('S' => 'a4'),
+                        'b' => array('S' => 'b4')
+                    ),
+                ),
+                array(
+                    array('a' => 'a4', 'b' => 'b4')
+                )
+            ),
+
+            // BatchGetItem results
+            array(
+                array(
+                    'Responses' => array(
+                        'foo' => array(
+                            array('a' => array('S' => 'a5'), 'b' => array('S' => 'b5')),
+                            array('a' => array('S' => 'a6'), 'b' => array('S' => 'b6')),
+                        ),
+                        'bar' => array(
+                            array('a' => array('S' => 'a7'), 'b' => array('S' => 'b7')),
+                            array('a' => array('S' => 'a8'), 'b' => array('S' => 'b8')),
+                        ),
+                    )
+                ),
+                array(
+                    array('a' => 'a5', 'b' => 'b5'),
+                    array('a' => 'a6', 'b' => 'b6'),
+                    array('a' => 'a7', 'b' => 'b7'),
+                    array('a' => 'a8', 'b' => 'b8'),
+                )
+            ),
+
+            // ListTables result
+            array(
+                array('TableNames' => array('foo', 'bar', 'baz')),
+                array()
+            ),
+        );
+    }
+
+    /**
+     * @dataProvider dataForGetItemsFromResultsTest
+     */
+    public function testCanGetItemsFromDifferentKindsOfResult(array $result, array $expectedItems)
+    {
+        /** @var \Aws\DynamoDb\DynamoDbClient $client */
+        $client = $this->getServiceBuilder()->get('dynamodb');
+        $result = new Model($result);
+
+        $actualItems = $client->getItemsFromResult($result)->toArray();
+        $this->assertEquals($expectedItems, $actualItems);
     }
 
     /**